+++ pg_bm25/src/lib.rs
@@ -20,7 +20,7 @@ extension_sql_file!("../sql/_bootstrap_quickstart.sql");
 #[pg_guard]
 pub unsafe extern "C" fn _PG_init() {
     index_access::options::init();
-    telemetry::telemetry::init();
+    telemetry::posthog::init();
 }
 
 /// This module is required by `cargo pgrx test` invocations.
