+++ pg_bm25/src/telemetry/mod.rs
@@ -1,1 +1,1 @@
-pub mod telemetry;
+pub mod posthog;
