+++ pg_bm25/src/telemetry/posthog.rs
