+++ pg_search/src/telemetry/posthog.rs
